cambia tipo de variable a float a campos precio habitacion y precio servicios

+++ database/migrations/2017_12_22_223244_create_precios_table.php
@@ -14,7 +14,7 @@ class CreatePreciosTable extends Migration
     {
         Schema::create('precios', function(Blueprint $table){
         $table->increments('id');
-        $table->integer('precio_habitacion');
+        $table->float('precio_habitacion',10,2);
         $table->integer('habitacion_id')->unsigned();
         $table->foreign('habitacion_id')->references('id')->on('habitaciones')->onDelete('cascade');
         $table->integer('tipo_moneda_id')->unsigned();
