+++ database/migrations/2017_12_22_244320_create_precios_servicio_table.php
@@ -14,7 +14,7 @@ class CreatePreciosServicioTable extends Migration
     {
         Schema::create('precios_servicio', function(Blueprint $table){
         $table->increments('id');
-        $table->integer('precio_servicio');
+        $table->float('precio_servicio',10,2);
         $table->integer('servicio_id')->unsigned();
         $table->foreign('servicio_id')->references('id')->on('servicios')->onDelete('cascade');
         $table->integer('tipo_moneda_id')->unsigned();
